Changed episode progress type from long to double to support deserializing non-integer values. Fixes #85

+++ src/DailyWire.Api.Middleware/Models/DwEpisodeDetails.cs
@@ -45,7 +45,7 @@ public class DwEpisodeDetails
     public double Duration { get; set; }
 
     [JsonPropertyName("progress")]
-    public long Progress { get; set; }
+    public double Progress { get; set; }
 
     [JsonPropertyName("status")]
     public DwStatus Status { get; set; }
